DT-727 Superclass to capture common gradle build test setup
