+++ src/test/kotlin/uk/gov/justice/digital/hmpps/gradle/functional/pluginmanagers/SpringBootPluginManagerTest.kt
@@ -1,13 +1,11 @@
 package uk.gov.justice.digital.hmpps.gradle.functional.pluginmanagers
 
 import org.assertj.core.api.Assertions.assertThat
-import org.gradle.internal.impldep.org.codehaus.plexus.util.FileUtils
 import org.gradle.testkit.runner.TaskOutcome
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.io.TempDir
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import uk.gov.justice.digital.hmpps.gradle.functional.GradleBuildTest
 import uk.gov.justice.digital.hmpps.gradle.functional.ProjectDetails
 import uk.gov.justice.digital.hmpps.gradle.functional.buildProject
 import uk.gov.justice.digital.hmpps.gradle.functional.buildProjectAndFail
@@ -15,30 +13,13 @@ import uk.gov.justice.digital.hmpps.gradle.functional.findJar
 import uk.gov.justice.digital.hmpps.gradle.functional.javaProjectDetails
 import uk.gov.justice.digital.hmpps.gradle.functional.kotlinProjectDetails
 import uk.gov.justice.digital.hmpps.gradle.functional.makeProject
-import java.io.File
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import java.util.jar.JarFile
 
-class SpringBootPluginManagerTest {
-
-  @AfterEach
-  fun `Delete project`() {
-    FileUtils.cleanDirectory(projectDir)
-  }
+class SpringBootPluginManagerTest : GradleBuildTest() {
 
   companion object {
-
-    @TempDir
-    @JvmStatic
-    lateinit var projectDir: File
-
-    @JvmStatic
-    fun defaultProjectDetails() = listOf(
-        Arguments.of(javaProjectDetails(projectDir)),
-        Arguments.of(kotlinProjectDetails(projectDir))
-    )
-
     @JvmStatic
     fun projectDetailsWithJunit4Tests() = listOf(
         Arguments.of(javaProjectDetails(projectDir).copy(testClass = javaJunit4Test())),
